Fix a cxx error and warning

+++ src/soundcard/sound.cxx
@@ -31,6 +31,8 @@
 #include <string>
 #include <vector>
 #include <sstream>
+#include <algorithm>
+#include <iterator>
 
 #include <cstdio>
 #include <cstdlib>
@@ -1214,7 +1216,7 @@ void SoundPort::init_stream(unsigned dir)
 			supported_rates[dir].push_back(*i);
 	ostringstream ss;
 	if (supported_rates[dir].size() > 1)
-		copy(supported_rates[dir].begin(), supported_rates[dir].end(), ostream_iterator<unsigned>(ss, " "));
+		copy(supported_rates[dir].begin(), supported_rates[dir].end(), ostream_iterator<double>(ss, " "));
 	else
 		ss << "Unknown";
